VocabVault-main/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
boto3==1.34.131
botocore==1.34.131
certifi==2024.6.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
deepl==1.18.0
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-filter==24.2
django-heroku==0.3.1
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
elevenlabs==1.3.1
future==1.0.0
geocoder==1.38.1
gunicorn==22.0.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==1.0.5
httpx==0.27.0
hyperframe==5.2.0
idna==2.10
jmespath==1.0.1
packaging==24.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
ratelim==0.1.6
requests==2.32.3
rfc3986==1.5.0
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
websockets==12.0
whitenoise==6.7.0
